requirements.txt
flask[async]
flask_cors
mysql-connector-python
redis[hiredis]
python-dotenv
